+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Delete duplicate C-Sytle functions for Cosmos SDK in wasm binding.
+- Add polling interval argument or function for setting the polling interval on event filters and pending transactions.
 
 ## [0.1.12] - 2022-04-25
 ### Added